회원가입, 로그인, mypage 추가

+++ ex/src/main/java/sogang/ip/ex/exchange/ExchangeRepository.java
@@ -2,6 +2,6 @@ package sogang.ip.ex.exchange;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExchangeRepository extends JpaRepository<Exchange, Long> {
+public interface ExchangeRepository extends JpaRepository<Exchange, Integer> {
 	
 }
